+++ src/pack_unpack.h
@@ -26,6 +26,6 @@ void Polp2BS(uint8_t bytes[SABER_POLYCOMPRESSEDBYTES],
              const uint16_t data[SABER_N]);
 int32_t Polp2BSCmp(const uint8_t bytes[SABER_POLYCOMPRESSEDBYTES],
                    const uint16_t data[SABER_N]);
-void PackSk(uint8_t bytes[SABER_SKPOLYBYTES], const uint16_t data[SABER_N]);
-void UnpackSk(const uint8_t *bytes, uint16_t data[SABER_N]);
+void PackSk(uint8_t bytes[SABER_SKPOLYBYTES], const uint8_t data[SABER_N]);
+void UnpackSk(const uint8_t *bytes, uint8_t data[SABER_N]);
 #endif
