+++ src/poly.h
@@ -5,9 +5,9 @@
 
 #include "SABER_params.h"
 
-void GenSecret(uint16_t s[SABER_L][SABER_N],
+void GenSecret(uint8_t s[SABER_L][SABER_N],
                const uint8_t seed[SABER_NOISE_SEEDBYTES]);
-void GenSInTime(uint16_t s[SABER_N], const uint8_t seed[SABER_NOISE_SEEDBYTES],
+void GenSInTime(uint8_t s[SABER_N], const uint8_t seed[SABER_NOISE_SEEDBYTES],
                 int32_t index);
 void GenSecretNTT(int32_t s[SABER_L][SABER_N],
                   const uint8_t seed[SABER_NOISE_SEEDBYTES]);
