+++ src/poly_mul.h
@@ -9,7 +9,7 @@ void PolyBaseMul(int32_t a[SABER_N], const int32_t b[SABER_N]);
 
 void PolyAdd(uint16_t res[SABER_N], int32_t in[SABER_N]);
 
-void PolyMulAcc(uint16_t a[2 * SABER_N], const uint16_t b[SABER_N],
+void PolyMulAcc(uint16_t a[2 * SABER_N], const uint8_t b[SABER_N],
                 uint16_t res[SABER_N]);
 void PolyMulAccFast(uint16_t a[SABER_N], const int32_t b[SABER_N],
                     uint16_t res[SABER_N]);
@@ -17,15 +17,14 @@ void MatrixVectorMulKP(const uint8_t *seed_a, const uint8_t *seed_s,
                        uint8_t sk[SABER_INDCPA_SECRETKEYBYTES],
                        uint16_t b[SABER_L][SABER_N]);
 #if defined(FASTGENA_SLOWMUL)
-void MatrixVectorMulEnc(const uint8_t *seed, uint16_t s[SABER_L][SABER_N],
+void MatrixVectorMulEnc(const uint8_t *seed, uint8_t s[SABER_L][SABER_N],
                         uint8_t *ciphertext);
-int32_t MatrixVectorMulEncCmp(const uint8_t *seed, uint16_t s[SABER_L][SABER_N],
+int32_t MatrixVectorMulEncCmp(const uint8_t *seed, uint8_t s[SABER_L][SABER_N],
                               const uint8_t *ciphertext);
-void InnerProdInTimeEnc(const uint8_t *bytes,
-                        const uint16_t s[SABER_L][SABER_N], uint8_t *ciphertext,
-                        const uint8_t m[SABER_KEYBYTES]);
+void InnerProdInTimeEnc(const uint8_t *bytes, const uint8_t s[SABER_L][SABER_N],
+                        uint8_t *ciphertext, const uint8_t m[SABER_KEYBYTES]);
 int32_t InnerProdInTimeEncCmp(const uint8_t *bytes,
-                              const uint16_t s[SABER_L][SABER_N],
+                              const uint8_t s[SABER_L][SABER_N],
                               const uint8_t *ciphertext,
                               const uint8_t m[SABER_KEYBYTES]);
 #elif defined(FASTGENA_FASTMUL) || defined(SLOWGENA_FASTMUL)
